refactor: email helper and data helper
Fixes rubocop warning for number of lines
